implement remove from cart functionality

+++ public/scripts.js
@@ -1,9 +1,12 @@
+const toCurrency = price => new Intl.NumberFormat('en-US', {
+    currency: 'USD',
+    style: 'currency'
+  }).format(price)
+
+
 document.addEventListener('DOMContentLoaded', () => {
   document.querySelectorAll('.price').forEach(node => {
-    node.textContent = new Intl.NumberFormat('en-US', {
-      currency: 'USD',
-      style: 'currency'
-    }).format(node.textContent);
+    node.textContent = toCurrency(node.textContent);
   });
   
   document.querySelectorAll('.date').forEach(node => {
@@ -14,13 +17,41 @@ document.addEventListener('DOMContentLoaded', () => {
   });
   
 
-  const cartDeleteBtns = document.querySelectorAll('.deleteFromCart');
+  const cartContent = document.querySelector('.cart-content');
+  let cartContentBody, cartTotalPrice;
 
-  cartDeleteBtns.forEach(c => c.addEventListener('click', (el) => {
-    fetch(`/cart/${el.currentTarget.dataset.id}/delete`, {
-      method: 'DELETE'
-    })
-      .then();
-  }))
+  if (cartContent) {
+    cartContentBody = cartContent.querySelector('tbody');
+    cartTotalPrice = cartContent.querySelector('.total-price');
+
+    cartContent.addEventListener('click', (e) => {
+      if (e.target.classList.contains('deleteFromCart')) {
+        fetch(`/cart/${e.target.dataset.id}/delete`, {
+          method: 'DELETE'
+        })
+          .then(res => res.json())
+          .then(cart => {
+            if (cart.courses.length) {
+              const tableBody = cart.courses.map(({title, price, count, id}) => {
+                return `
+                <tr>
+                  <td>${title}</td>
+                  <td class="price">${price}</td>
+                  <td>${count}</td>
+                  <td>
+                    <button class="uk-button uk-button-danger deleteFromCart" data-id="${id}">Delete</button>
+                  </td>
+                </tr>`
+              }).join('');
+    
+              cartContentBody.innerHTML = tableBody;
+              cartTotalPrice.textContent = toCurrency(cart.price);
+            } else {
+              cartContent.innerHTML = '<p>Your cart is empty!</p>'
+            }
+          });
+      }
+    });
+  }
 });
